refactor: move into groups directory
