+++ components/pit/groups/Groups.tsx
@@ -1,6 +1,6 @@
-import {useFetchGroups} from "../../src/features/groups/hook";
+import {useFetchGroups} from "../../../src/features/groups/hook";
 import {GroupList} from "./GroupList";
-import styles from "../../styles/Pit.module.scss";
+import styles from "../../../styles/Pit.module.scss";
 import {GroupForm} from "./GroupForm";
 import {Button} from "@mui/material";
 import {useState} from "react";
